feat: Implement study export and import functionality
- Added StudyExport class to handle the conversion of Study objects to LLM-friendly JSON schemas.
- Introduced StudyExportContext for managing relationships during the export process.
- Created StudyImport class to reconstruct Study objects from JSON schemas, maintaining compatibility with existing studies.
- Added StudyImportContext to track relationships during the import process.
- Implemented methods for exporting study metadata, schedules, observations, interventions, and consent.
- Set up methods for importing study metadata, schedules, screening forms, observations, interventions, and consent.
