+++ designer_v2/lib/domain/study_import.dart
@@ -1,0 +1,78 @@
+import 'package:studyu_core/core.dart';
+import 'package:studyu_designer_v2/domain/study_import_context.dart';
+
+/// Handles the import of Study objects from LLM-friendly JSON schemas.
+/// 
+/// This class contains all the logic for converting simplified JSON
+/// representations back into full StudyU Study objects, maintaining
+/// backward compatibility with native UI-created studies.
+class StudyImport {
+  StudyImport._();
+
+  /// Creates a [Study] object from a simplified JSON schema.
+  /// 
+  /// Reconstructs a complete Study object from the LLM-friendly schema format.
+  /// The import process resolves natural language references back to internal
+  /// IDs and creates all necessary Study components.
+  /// 
+  /// Parameters:
+  /// - [schema]: The JSON schema containing study data
+  /// - [ownerId]: The ID of the user who will own this study
+  /// 
+  /// Returns a fully configured [Study] object ready for use in the StudyU platform.
+  static Study fromSchema(
+    Map<String, dynamic> schema, {
+    required String ownerId,
+  }) {
+    final study = Study.withId(ownerId);
+    final context = StudyImportContext();
+    
+    _importMetadata(schema['metadata'] as Map<String, dynamic>? ?? {}, study);
+    _importStudySchedule(
+      schema['studySchedule'] as Map<String, dynamic>?,
+      study.schedule,
+    );
+    _importScreening(schema['screening'], study, context);
+    _importObservations(schema['observations'], study, context);
+    _importInterventions(schema['interventions'], study);
+    _importConsent(schema['consent'], study);
+    
+    return study;
+  }
+
+  // TODO: Add all import methods from simplified_study_converter.dart
+  static void _importMetadata(Map<String, dynamic> metadata, Study study) {
+    // Implementation will be copied from original file
+  }
+
+  static void _importStudySchedule(
+    Map<String, dynamic>? data,
+    StudySchedule schedule,
+  ) {
+    // Implementation will be copied from original file
+  }
+
+  static void _importScreening(
+    dynamic screeningJson,
+    Study study,
+    StudyImportContext context,
+  ) {
+    // Implementation will be copied from original file
+  }
+
+  static void _importObservations(
+    dynamic observationsJson,
+    Study study,
+    StudyImportContext context,
+  ) {
+    // Implementation will be copied from original file
+  }
+
+  static void _importInterventions(dynamic data, Study study) {
+    // Implementation will be copied from original file
+  }
+
+  static void _importConsent(dynamic data, Study study) {
+    // Implementation will be copied from original file
+  }
+}
